Button symptom tracker (#222)
