+++ client/src/components/Layout.js
@@ -26,12 +26,13 @@ const theme = createMuiTheme({
 });
 const useStyles = makeStyles(theme => ({
   //the styles goes here as an object
-    root: {
-        textAlign: 'center',
-  backgroundImage: "linear-gradient(#d7e1fa, #bbcef9)",
-  height: "100vh"
-        
-    },
+  root: {
+    textAlign: 'center',
+    backgroundImage: "linear-gradient(#d7e1fa, #bbcef9)",
+    overflowY: 'scroll',
+    height: "85vh"
+
+  },
   logo: {
     width: '75px',
 
@@ -49,16 +50,9 @@ const useStyles = makeStyles(theme => ({
     },
   },
 }));
-const dateOptions = {
-  weekday: 'long',
-  year: 'numeric',
-  month: 'long',
-  day: 'numeric',
-};
 const Layout = (props) => {
   const classes = useStyles();
-  const { handleSignOut, userSession, authed } = props;
-  const today = new Date();
+  const { handleSignOut, authed } = props;
 
   return (
     <MuiThemeProvider theme={theme}>
@@ -71,20 +65,13 @@ const Layout = (props) => {
             <Button size="medium" color="secondary" variant="contained" onClick={handleSignOut}>
               Sign Out
             </Button>
-            <h4>
-              Hello, <b>{userSession.loadUserData().profile.name} </b>
-            </h4>
-            <h5>
-              Today is <b>{today.toLocaleDateString(undefined, dateOptions)}</b>{' '}
-            </h5>
-            <hr className="hr" />
             {props.children}
 
             <NavBar />
           </div>
         ) : (
-          <Login />
-        )}
+            <Login />
+          )}
       </CssBaseline>
     </MuiThemeProvider>
   );
