+++ client/src/components/MyHealthLog.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import Button from '@material-ui/core/Button';
+
+export default function MyHealthLog() {
+  return (
+    <Button color="secondary"variant='contained' href='/symptomsurvey'>Take a survey</Button>
+  )
+}
